src/document_intelligence/requirements.txt
# Azure Functions Core
azure-functions==1.18.0

# Azure SDK - Document Intelligence
azure-ai-documentintelligence==1.0.0b1
azure-core==1.29.5

# Azure SDK - Storage
azure-storage-blob==12.19.0
azure-identity==1.15.0

# HTTP and Async
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Data Validation and Serialization
pydantic==2.5.2
pydantic-core==2.14.5

# Logging and Monitoring
applicationinsights==0.11.10
opencensus-ext-azure==1.1.13
structlog==23.2.0

# Utilities
python-dotenv==1.0.0
typing-extensions==4.8.0

# Development and Testing (optional in production)
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-mock==3.12.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1